App: restructure and clean the code
