+++ src/components/app/App.js
@@ -1,4 +1,4 @@
-import logo from './logo.svg';
+import logo from '../../imgs/react.svg';
 import './App.css';
 
 function App() {
